delete food-list.tsx
